+++ plugin_ide.core.tests/src/mmrnmhrm/core/search/DeeSearchEngine_BasicTest.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 
 import melnorme.lang.tooling.context.ISemanticContext;
-import melnorme.lang.tooling.engine.NotFoundErrorElement;
+import melnorme.lang.tooling.engine.ErrorElement;
 import melnorme.lang.tooling.symbols.INamedElement;
 import mmrnmhrm.core.DLTKUtils;
 import mmrnmhrm.core.engine_client.DToolClient;
@@ -168,7 +168,7 @@ public class DeeSearchEngine_BasicTest extends DeeSearchEngine_Test {
 		ISemanticContext mr = DToolClient.getDefault().getResolvedModule(filePath).getSemanticContext();
 		DefinitionVariable defVar = assertInstance(defUnit, DefinitionVariable.class);
 		INamedElement target = defVar.type.findTargetDefElement(mr);
-		assertTrue(target == null || target instanceof NotFoundErrorElement);
+		assertTrue(target == null || target instanceof ErrorElement);
 	}
 	
 }