+++ plugin_tooling/src-test/dtool/engine/analysis/NamedElement_CommonTest.java
@@ -21,7 +21,7 @@ import melnorme.lang.tooling.ast.util.NodeElementUtil;
 import melnorme.lang.tooling.context.EmptySemanticResolution;
 import melnorme.lang.tooling.context.ISemanticContext;
 import melnorme.lang.tooling.engine.NotAValueErrorElement;
-import melnorme.lang.tooling.engine.NotFoundErrorElement;
+import melnorme.lang.tooling.engine.ErrorElement;
 import melnorme.lang.tooling.engine.PickedElement;
 import melnorme.lang.tooling.engine.completion.CompletionScopeLookup;
 import melnorme.lang.tooling.engine.resolver.NamedElementSemantics;
@@ -114,8 +114,8 @@ public abstract class NamedElement_CommonTest extends CommonNodeSemanticsTest {
 		NamedElementSemantics semantics = namedElement.getSemantics(context);
 		IConcreteNamedElement concreteElement = semantics.resolveConcreteElement().result;
 		
-		if(concreteElement instanceof NotFoundErrorElement) {
-			NotFoundErrorElement notFoundError = (NotFoundErrorElement) concreteElement;
+		if(concreteElement instanceof ErrorElement) {
+			ErrorElement notFoundError = (ErrorElement) concreteElement;
 			assertTrue(notFoundError.getModulePath() == namedElement.getModulePath());
 			assertTrue(notFoundError.getParent() != null);
 			assertTrue(NodeElementUtil.isContainedIn(notFoundError.getParent(), namedElement));