+++ plugin_tooling/src-test/dtool/engine/analysis/NamedElements_Test.java
@@ -18,7 +18,7 @@ import java.util.concurrent.ExecutionException;
 import melnorme.lang.tooling.ast.ASTNodeFinder;
 import melnorme.lang.tooling.ast_actual.ASTNode;
 import melnorme.lang.tooling.context.ISemanticContext;
-import melnorme.lang.tooling.engine.NotFoundErrorElement;
+import melnorme.lang.tooling.engine.ErrorElement;
 import melnorme.lang.tooling.engine.PickedElement;
 import melnorme.lang.tooling.engine.intrinsics.CommonLanguageIntrinsics.IntrinsicProperty;
 import melnorme.lang.tooling.engine.intrinsics.CommonLanguageIntrinsics.IntrinsicProperty2;
@@ -126,8 +126,8 @@ public class NamedElements_Test extends NamedElement_CommonTest {
 			
 			visitConcrete(parseDefUnit("template xxx() { }"));
 			visitConcrete(parseDefUnit("template blah(int xxx) { }"));
-			visitAliasElement(parseDefUnit("template blah(xxx) { }"), NotFoundErrorElement.NOT_FOUND__NAME);
-			visitAliasElement(parseDefUnit("template blah(alias xxx) { }"), NotFoundErrorElement.NOT_FOUND__NAME);
+			visitAliasElement(parseDefUnit("template blah(xxx) { }"), ErrorElement.NOT_FOUND__NAME);
+			visitAliasElement(parseDefUnit("template blah(alias xxx) { }"), ErrorElement.NOT_FOUND__NAME);
 			visitConcrete(parseDefUnit("template blah(this xxx) { }"));
 			
 			visitConcrete(parseDefUnit("mixin blah xxx;"));