+++ plugin_tooling/src-test/dtool/resolver/DefUnitResultsChecker.java
@@ -22,7 +22,7 @@ import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.Map;
 
-import melnorme.lang.tooling.engine.NotFoundErrorElement;
+import melnorme.lang.tooling.engine.ErrorElement;
 import melnorme.lang.tooling.engine.intrinsics.CommonLanguageIntrinsics.IPrimitiveDefUnit;
 import melnorme.lang.tooling.symbols.INamedElement;
 import melnorme.utilbox.core.fntypes.Function;
@@ -69,7 +69,7 @@ public class DefUnitResultsChecker extends CommonTestUtils {
 				iterator.remove();
 			} else if(ignoreIntrinsics && defElement.isLanguageIntrinsic()) {
 				iterator.remove();
-			} else if(defElement instanceof NotFoundErrorElement) {
+			} else if(defElement instanceof ErrorElement) {
 				iterator.remove();
 			}
 		}